Work In Progress 1) Add gps functions to MSL output 2) Put comms on a separate thread so we can parallelise calc of current data and fetch of next 3) Back off fetch rate (default 10Hz) if we can't keep up)

+++ src/bsh/commands/timeNow.java
@@ -6,8 +6,8 @@ import bsh.Interpreter;
 public class timeNow
 {
     private static long startTicks = System.currentTimeMillis();
-    public static int invoke(Interpreter env, CallStack callstack)
+    public static double invoke(Interpreter env, CallStack callstack)
     {
-        return (int) ((System.currentTimeMillis() - startTicks)/1000);
+        return (double) ((System.currentTimeMillis() - startTicks)/1000.0);
     }
 }
